chore: add parent PID to coder ssh log file name
Part of bringing coder ssh to parity with coder vscodessh is associating
the log files with a particular parent process (in this case, the ssh
process that spawned the coder CLI via ProxyCommand). coder vscodessh
named log files using the parent PID, but coder ssh is missing this. Add
the parent PID to the log file name when used in stdio mode so that
the VS Code extension will be able to identify the correct log file.